fix(all): fix ci build
