--- packages/core/src/ui/components/ActivityDetail/ActivityType.ts
@@ -1,11 +1,0 @@
-export type TransactionActivityType =
-  | 'outgoing'
-  | 'incoming'
-  | 'delegation'
-  | 'delegationRegistration'
-  | 'delegationDeregistration'
-  | 'self';
-
-export type RewardsActivityType = 'rewards';
-
-export type ActivityType = TransactionActivityType | RewardsActivityType;
