--- packages/core/src/ui/components/ActivityDetail/RewardsInfo.ts
@@ -1,10 +1,0 @@
-type RewardItem = {
-  pool?: { name: string; ticker: string; id: string };
-  amount: string;
-};
-
-export type RewardsInfo = {
-  totalAmount: string;
-  spendableEpoch: number;
-  rewards: RewardItem[];
-};
