+++ packages/core/src/ui/components/ActivityDetail/index.ts
@@ -1,7 +1,5 @@
 export * from './TransactionDetails';
 export * from './RewardsDetails';
-export * from './activityType';
 export * from './ActivityTypeIcon';
 export * from './TransactionDetailAsset';
 export * from './TransactionInputOutput';
-export * from './rewardsInfo';
